Move sequence util and fix merge conflict
