--- keras/datasets/text_utils.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-def _remove_long_seq(maxlen, seq, label):
-    new_seq, new_label = [], []
-    for x, y in zip(seq, label):
-        if len(x) < maxlen:
-            new_seq.append(x)
-            new_label.append(y)
-    return new_seq, new_label
