+++ keras/preprocessing/sequence.py
@@ -191,3 +191,22 @@ def skipgrams(sequence, vocabulary_size,
         random.shuffle(labels)
 
     return couples, labels
+
+
+def _remove_long_seq(maxlen, seq, label):
+    """Removes sequences that exceed the maximum length.
+
+    # Arguments
+        maxlen: int, maximum length
+        seq: list of lists where each sublist is a sequence
+        label: list where each element is an integer
+
+    # Returns
+        new_seq, new_label: shortened lists for `seq` and `label`.
+    """
+    new_seq, new_label = [], []
+    for x, y in zip(seq, label):
+        if len(x) < maxlen:
+            new_seq.append(x)
+            new_label.append(y)
+    return new_seq, new_label
