+++ tests/keras/preprocessing/image_test.py
@@ -29,14 +29,6 @@ class TestImage:
     def teardown_class(cls):
         del cls.all_test_images
 
-    def test_random_transforms(self):
-        x = np.random.random((2, 28, 28))
-        assert image.random_rotation(x, 45).shape == (2, 28, 28)
-        assert image.random_shift(x, 1, 1).shape == (2, 28, 28)
-        assert image.random_shear(x, 20).shape == (2, 28, 28)
-        assert image.random_zoom(x, (5, 5)).shape == (2, 28, 28)
-        assert image.random_channel_shift(x, 20).shape == (2, 28, 28)
-
     def test_image_data_generator(self):
         for test_images in self.all_test_images:
             img_list = []
@@ -181,13 +173,6 @@ class TestImage:
                 im.save(os.path.join(tmp_folder, filename))
                 count += 1
 
-        for class_mode in ['sparse', 'binary', 'categorical']:
-            generator = image.ImageDataGenerator(data_format='channels_first')
-            dir_iterator = generator.flow_from_directory(tmp_folder, class_mode=class_mode)
-            batch = next(dir_iterator)
-            # check if input and output have the same number of samples
-            assert batch[0].shape[0] == batch[1].shape[0]
-
         # create iterator
         generator = image.ImageDataGenerator()
         dir_iterator = generator.flow_from_directory(tmp_folder, class_mode='input')
@@ -247,6 +232,14 @@ class TestImage:
         with pytest.raises(ValueError):
             x = np.random.random((height, width, 5, 3))  # neither RGB nor gray-scale
             img = image.img_to_array(x, data_format='channels_last')
+
+    def test_random_transforms(self):
+        x = np.random.random((2, 28, 28))
+        assert image.random_rotation(x, 45).shape == (2, 28, 28)
+        assert image.random_shift(x, 1, 1).shape == (2, 28, 28)
+        assert image.random_shear(x, 20).shape == (2, 28, 28)
+        assert image.random_zoom(x, (5, 5)).shape == (2, 28, 28)
+        assert image.random_channel_shift(x, 20).shape == (2, 28, 28)
 
 
 if __name__ == '__main__':
